add S3File.cpp
